Rename schemas/documents -> document
