+++ imports/client/components/PuzzlePage.tsx
@@ -35,7 +35,7 @@ import Puzzles from '../../lib/models/puzzles';
 import Tags from '../../lib/models/tags';
 import { userMayWritePuzzlesForHunt } from '../../lib/permission_stubs';
 import { ChatMessageType } from '../../lib/schemas/chat';
-import { DocumentType } from '../../lib/schemas/documents';
+import { DocumentType } from '../../lib/schemas/document';
 import { GuessType } from '../../lib/schemas/guess';
 import { PuzzleType } from '../../lib/schemas/puzzles';
 import { TagType } from '../../lib/schemas/tags';
